Add files via upload
it is initial version of software. it has only a Structure and print in some steps to show the flow. 
now a function named "GoForward" is called in main and in this function a random number is called and this number is the function which should be run as Event Card.